data-worker/requirements.txt
yfinance==0.2.36
kafka-python==2.0.2